Moves tests to the correct package to reflect coverage
