+++ opensrp-reporting/src/test/java/org/smartregister/reporting/util/AggregationUtilTest.java
@@ -1,4 +1,4 @@
-package org.smartregister.reporting.utils;
+package org.smartregister.reporting.util;
 
 import org.junit.Assert;
 import org.junit.Test;
